requirements.txt
openai==1.73.0
supabase==2.15.0
python-dotenv==1.1.0
requests==2.32.3
beautifulsoup4==4.13.3